requirements.txt
Flask
psycopg2-binary
pywebview
openpyxl
Werkzeug
